Set unicode literals everywhere
This is further work toward Python3 suppot.
Replace isinstance(..., str) with isinstance(..., six.string_types) in a bunch of places, mostly in tests.
Strings in __all__ in Py2 must be bytestrings so explicitly declare them as such.
The odd change in interface.py is due to element tree incorrectly returning a unicode string when the encoding string is itself unicode.