Return meta information in tuple format
The related_events endpoint returns two result sets for add_on events
and upsell events, so the meta information for each is returned in tuple
format. Additionally the Paginator Mixin is updated to support
specifying the location within the response that the pagination
information can be found (rather than assuming it will always be in a
dictionary with key 'result').