handle order unavailable on reservation success
This is an edge case where a trolley is built using `get_trolley`
then the user attempts to add something additional to the trolley at
reservation time, the API is unable to find any available product that
matches the new addition, but continues to reserve the rest of the
trolley. In this situation if user was using the
`raise_on_unavailable_order` flag on `make_reservation` then the details
of the reserved orders would be lost.

This change attaches the reservation and meta data to the exception so
that the user is aware of them.